// types/enums.ts
export enum GameStatus {
  WANT_TO_PLAY = 'want_to_play',
  CURRENTLY_PLAYING = 'currently_playing',
  COMPLETED = 'completed',
  ON_HOLD = 'on_hold',
  DROPPED = 'dropped'
}

export enum Platform {
  PC = 'pc',
  PLAYSTATION_5 = 'ps5',
  PLAYSTATION_4 = 'ps4',
  XBOX_SERIES = 'xbox_series',
  XBOX_ONE = 'xbox_one',
  NINTENDO_SWITCH = 'nintendo_switch',
  MOBILE = 'mobile',
  VR = 'vr'
}

export enum Genre {
  ACTION = 'action',
  ADVENTURE = 'adventure',
  RPG = 'rpg',
  STRATEGY = 'strategy',
  SIMULATION = 'simulation',
  SPORTS = 'sports',
  RACING = 'racing',
  FIGHTING = 'fighting',
  SHOOTER = 'shooter',
  PUZZLE = 'puzzle',
  HORROR = 'horror',
  PLATFORMER = 'platformer',
  INDIE = 'indie'
} 